Create generateToken file in utils
