+++ server/api/meeting/index.post.ts
@@ -1,28 +1,6 @@
 import { AbortError, ServiceError, fql } from 'fauna'
 import type { MeetingFields } from '~/types'
-import jwt from 'jsonwebtoken'
-import { v4 as uuidv4 } from 'uuid'
-
-// Function to generate a management token
-function generateManagementToken() {
-	const payload = {
-		access_key: process.env.APP_ACCESS_KEY,
-		type: 'management',
-		version: 2,
-		iat: Math.floor(Date.now() / 1000),
-		nbf: Math.floor(Date.now() / 1000),
-	}
-
-	return new Promise((resolve, reject) => {
-		jwt.sign(
-			payload,
-			process.env.APP_SECRET!,
-			{ algorithm: 'HS256', expiresIn: '24h', jwtid: uuidv4() },
-			(err, token) => (err ? reject(err) : resolve(token)),
-		)
-	})
-}
-
+import { generateManagementToken } from '~/server/utils/generateToken'
 // Endpoint for creating a meeting
 export default defineEventHandler(async (event) => {
 	// Initialize Fauna client
